Resolving conflict
