+++ src/_convert_eight_bit_list_to_string.py
